Fix up the compression/offline compression handling.
Offline compression needs to be disabled for devstack to work, but
it needs to be enabled for the RDO-style packaging to work. I had
originally just disabled it, but now that it's re-enabled (and the
handling of the re-enabling is done a bit more robustly) it seems
to be working okay.